Put the user definition in a separate file
